Changed amount data type in CashFlowEntity.kt, Created dropdown_category_items.xml for category dropdown, Changed context in fragment_home.xml, Created NewCashFlow.kt and fragment_new_cash_flow.xml, Cleaned up History.kt, Stats.kt, Home.kt, Relocated Home.kt, Created NewCashFlowViewModel.kt, Added few string and a string-array in strings.xml.

+++ app/src/main/java/com/github/sirdeerhead/dailyspending/room/CashFlowEntity.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 data class CashFlowEntity(
     @PrimaryKey(autoGenerate = true) val id: Int = 0,
     val date: String = "",
-    val amount: Int = 0,
+    val amount: Double = 0.0,
     val category: String = "",
     val description: String = "",
 )
